+++ washee_app/lib/features/booking/domain/usecases/delete_booking.dart
@@ -1,7 +1,8 @@
 import 'package:equatable/equatable.dart';
 import 'package:washee/core/standards/base_usecase/usecase.dart';
+import 'package:washee/features/booking/data/repositories/book_repository_impl.dart';
 
-import '../repositories/book_repository.dart';
+
 
 class DeleteBookingUseCase implements UseCase<bool, DeleteBookingParams> {
   final BookRepository repository;
