+++ washee_app/lib/features/booking/domain/usecases/get_bookings.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:washee/core/standards/base_usecase/usecase.dart';
 import 'package:washee/features/booking/data/models/booking_model.dart';
-import 'package:washee/features/booking/domain/repositories/book_repository.dart';
+import 'package:washee/features/booking/data/repositories/book_repository_impl.dart';
 
 class GetBookingsUseCase implements UseCase<List<BookingModel>, GetBookingsParams> {
   final BookRepository repository;
