+++ washee_app/lib/features/booking/domain/usecases/has_current_booking.dart
@@ -1,7 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:washee/core/standards/base_usecase/usecase.dart';
-
-import '../repositories/book_repository.dart';
+import 'package:washee/features/booking/data/repositories/book_repository_impl.dart';
 
 class HasCurrentBookingUseCase implements UseCase<bool, HasCurrentBookingParams> {
   final BookRepository repository;
