--- washee_app/lib/features/location/domain/repositories/connect_repository.dart
@@ -1,5 +1,0 @@
-import 'package:dio/dio.dart';
-
-abstract class ConnectRepository {
-  Future<Response> connect();
-}
