--- washee_app/lib/features/location/domain/repositories/get_machines_repository.dart
@@ -1,5 +1,0 @@
-import 'package:washee/features/location/data/models/box_machine_model.dart';
-
-abstract class GetMachinesRepository {
-  Future<List<MachineModel>> getMachines();
-}
