--- washee_app/lib/features/location/domain/repositories/unlock_repository.dart
@@ -1,8 +1,0 @@
-import 'package:washee/features/location/data/models/box_machine_model.dart';
-
-abstract class UnlockRepository {
-  Future<MachineModel?> unlock(MachineModel machine, Duration duration);
-  Future<bool> getWifiPermission();
-  Future<bool> connectToBoxWifi();
-  Future<bool> disconnectFromBoxWifi();
-}
