+++ washee_app/lib/features/location/domain/usecases/connect.dart
@@ -1,7 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:washee/core/standards/base_usecase/usecase.dart';
-
-import '../repositories/connect_repository.dart';
+import 'package:washee/features/location/data/repositories/connect_repo.dart';
 
 class ConnectUseCase implements UseCase<Response, NoParams> {
   final ConnectRepository repository;
