+++ washee_app/lib/features/location/domain/usecases/get_wifi_permission.dart
@@ -1,6 +1,5 @@
 import 'package:washee/core/standards/base_usecase/usecase.dart';
-
-import '../repositories/unlock_repository.dart';
+import 'package:washee/features/location/data/repositories/unlock_repo.dart';
 
 class GetWifiPermissionUsecase implements UseCase<bool, NoParams> {
   final UnlockRepository repository;
