+++ washee_app/lib/features/location/domain/usecases/unlock.dart
@@ -4,10 +4,9 @@ import 'package:washee/core/externalities/web/web_connector.dart';
 import 'package:washee/core/standards/time/date_helper.dart';
 import 'package:washee/core/standards/base_usecase/usecase.dart';
 import 'package:washee/features/location/data/models/box_machine_model.dart';
+import 'package:washee/features/location/data/repositories/unlock_repo.dart';
 import 'package:washee/features/location/domain/usecases/disconnect_box_wifi.dart';
 import 'package:washee/injection_container.dart';
-
-import '../../../location/domain/repositories/unlock_repository.dart';
 
 class UnlockUseCase implements UseCase<MachineModel?, UnlockParams> {
   final UnlockRepository repository;
