+++ washee_app/test/features/account/data/datasources/web_account_remote_test.dart
@@ -4,7 +4,7 @@ import 'package:dio/dio.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:washee/core/externalities/web/web_connector.dart';
-import 'package:washee/features/account/data/datasources/account_remote.dart';
+import 'package:washee/features/account/data/datasources/web_account_remote.dart';
 
 import '../../../../fixtures/entities/account/accounts.dart';
