+++ washee_app/test/features/account/data/repositories/account_repository_test.dart
@@ -2,7 +2,7 @@ import 'dart:core';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:washee/features/account/data/datasources/account_remote.dart';
+import 'package:washee/features/account/data/datasources/web_account_remote.dart';
 import 'package:washee/features/account/data/models/web_account_model.dart';
 import 'package:washee/features/account/data/repositories/account_repository_impl.dart';
 import 'package:washee/features/account/domain/entities/account_entity.dart';
