+++ washee_app/test/features/account/data/repositories/user_repository_test.dart
@@ -2,7 +2,7 @@ import 'dart:core';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:washee/features/account/data/datasources/user_remote.dart';
+import 'package:washee/features/account/data/datasources/web_user_remote.dart';
 import 'package:washee/features/account/data/models/web_user_model.dart';
 import 'package:washee/features/account/data/repositories/user_repository.dart';
 import 'package:washee/features/account/domain/entities/user_entity.dart';
