+++ washee_app/test/features/booking/data/repositories/book_repository_impl_test.dart
@@ -3,13 +3,8 @@ import 'dart:core';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:washee/core/standards/time/date_helper.dart';
-import 'package:washee/core/externalities/network/network_info.dart';
 import 'package:washee/features/booking/data/datasources/book_remote.dart';
-import 'package:washee/features/booking/data/repositories/book_repository_impl.dart';
 import 'package:timezone/timezone.dart' as tz;
-import 'package:washee/features/booking/domain/repositories/book_repository.dart';
-
-class MockNetworkInfo extends Mock implements NetworkInfo {}
 
 class MockBookRemote extends Mock implements BookRemote {}
 
@@ -20,28 +15,20 @@ class FakeLocation extends Mock implements tz.Location {}
 // class MockBookingModel extends Mock implements BookingModelHelper {}
 
 void main() {
-  late BookRepository sut_bookRepositoryImpl;
-  late MockNetworkInfo mockNetworkInfo;
-  late MockBookRemote mockRemote;
-  // late MockBookingModel mockBookingHelper;
-  //late MockDateHelper mockDateHelper;
-  //late BookingModel mockBookingModelResult;
-  late Map<String, dynamic> mockBookingJson;
-  //late DateTime currentTime;
 
   setUp(() {
+    /*
     registerFallbackValue(FakeLocation());
-    mockRemote = MockBookRemote();
-    mockNetworkInfo = MockNetworkInfo();
+    var mockRemote = MockBookRemote();
     // mockBookingHelper = MockBookingModel();
     //mockDateHelper = MockDateHelper();
-    sut_bookRepositoryImpl =
-        BookRepositoryImpl(networkInfo: mockNetworkInfo, remote: mockRemote );
+    var sut_bookRepository =
+        BookRepository(remote: mockRemote);
     //currentTime = DateTime.now();
     var booking1_start_time = DateTime(2022, 01, 01, 2, 0);
     var booking1_end_time =
         booking1_start_time.add(Duration(hours: 2, minutes: 30));
-    mockBookingJson = {
+    var mockBookingJson = {
       "bookingID": 12,
       "startTime": booking1_start_time,
       "endTime": booking1_end_time,
@@ -49,6 +36,7 @@ void main() {
       "serviceResource": "https://mocked_serviceResource/1",
       "accountResource": "https://mocked_accountResource/1"
     };
+    */
 
     //mockBookingModelResult = BookingModel(
     //    bookingID: 12,
@@ -107,12 +95,11 @@ void main() {
 //         serviceResource: mockBookingJson["serviceResource"],
 //         accountResource: mockBookingJson["accountResource"])).called(1);
 //   });
-
+/*
   test(
       'should verify that remote.postBooking() is never called when isConnected is false',
       () async {
     //arrange
-    when(() => mockNetworkInfo.isConnected).thenAnswer((_) async => false);
     when(() => mockRemote.postBooking(
             startTime: mockBookingJson["startTime"],
             machineResource: mockBookingJson["machineResource"],
@@ -133,4 +120,5 @@ void main() {
         serviceResource: mockBookingJson["serviceResource"],
         accountResource: mockBookingJson["accountResource"]));
   });
+  */
 }
