+++ washee_app/test/features/booking/domain/usecases/post_booking_test.dart
@@ -3,7 +3,7 @@ import 'dart:core';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:washee/features/booking/data/models/booking_model.dart';
-import 'package:washee/features/booking/domain/repositories/book_repository.dart';
+import 'package:washee/features/booking/data/repositories/book_repository_impl.dart';
 import 'package:washee/features/booking/domain/usecases/post_booking.dart';
 
 class MockBookRepository extends Mock implements BookRepository {}
